manage notes by user_id

+++ src/modules/v1/controllers/notes/index.ts
@@ -11,10 +11,11 @@ export default class NotesController extends ControllerV1 {
   }
 
   public async getNotes(ctx: Context) {
+    const { _id } = ctx.state.auth;
     try {
       const data = await this._notesModel
         .getCollection()
-        .find()
+        .find({ user_id: _id })
         .sort([['updated_at', -1]])
         .lean();
       return this._api.pass(ctx, data);
@@ -25,8 +26,13 @@ export default class NotesController extends ControllerV1 {
 
   public async getNoteById(ctx: Context) {
     const { id } = ctx.params;
+    const { _id: userId } = ctx.state.auth;
+
     try {
-      const data = await this._notesModel.getCollection().findById(id).lean();
+      const data = await this._notesModel
+        .getCollection()
+        .findOne({ _id: id, user_id: userId })
+        .lean();
       return this._api.pass(ctx, data);
     } catch (err: any) {
       return this._api.fail(ctx, Config.codes.ERR_00000, err.message);
@@ -35,8 +41,9 @@ export default class NotesController extends ControllerV1 {
 
   public async updateNote(ctx: Context) {
     const { _id } = ctx.request.body;
+    const { _id: userId } = ctx.state.auth;
     const updateData = Object.assign(
-      { updated_at: Date.now() },
+      { updated_at: Date.now(), user_id: userId },
       ctx.request.body,
     );
     try {
@@ -51,8 +58,10 @@ export default class NotesController extends ControllerV1 {
 
   public async createNote(ctx: Context) {
     const { body } = ctx.request;
+    const { _id: userId } = ctx.state.auth;
+    const createData = Object.assign({ user_id: userId }, body);
     try {
-      const data = await this._notesModel.getCollection().create(body);
+      const data = await this._notesModel.getCollection().create(createData);
       return this._api.pass(ctx, data);
     } catch (err: any) {
       return this._api.fail(ctx, Config.codes.ERR_00000, err.message);
@@ -61,10 +70,11 @@ export default class NotesController extends ControllerV1 {
 
   public async deleteNote(ctx: Context) {
     const { id } = ctx.params;
+    const { _id: userId } = ctx.state.auth;
     try {
       const data = await this._notesModel
         .getCollection()
-        .deleteOne({ _id: id });
+        .deleteOne({ _id: id, user_id: userId });
       return this._api.pass(ctx, data);
     } catch (err: any) {
       return this._api.fail(ctx, Config.codes.ERR_00000, err.message);
